feat(meeting): better heading for take aways
Because of how `org-capture` works, there were times that it wasn't easy
to tell the context of a todo item that ended up in the Take Aways
section. This repeats the filename in the Take Aways header so it also
show up in your agenda.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>