fix(meeting): avoid path separators in filenames
Previously I'd have a lot of meetings that I'd call "BLAH AI/ML BLAH",
but this caused problems because of th path separator. This new function
leaves the path separator in the title, but doesn't save the files with
the path sepator in the filename.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>