fix(save): another try at improving saves
The fundamental challenge is that I want to save the file and add in
org-id values at the same time. This doesn't work well, so instead we
save the file and then add in the headers, and save the file again. This
logic makes it so the file is only saved the first time if it doesn't
already exist on the filesystem.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>