fix(meeting): fix parsing of meeting names
I thought that emacs had a variable that was the platform specific path
delimiter, but it doesn't appear that `f-path-separator` exists
everywhere. I just swapped it out for `/`. So now, any meeting that has
a `/` in the title will have the `/` character removed from the
filename. This is really helpful as I have a lot of meetings that have
"AI/ML" in the meeting name.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>