fix(save): fixes infinite loop in saving
I tried to find good ways to do this in emacs, but in the end, the best
way to have it insert IDs into each headline in an org file is to save
it on the second time that the file has been saved. Otherwise it causes
problems with entering the infromation into the `org-roam.db` file.

Lame, but files are usually saved often.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>