feat(company): adds `org-insert-company`
This brings in support for a new object type, a company. This can be
really useful when you're tracking information about who works for what
company and makes it so you can jump around and find people who work for
a company.

It also adds in a little bit of a fix for inserting dates into the
skeleton templates and simplifies `org-insert-person` into a more
generic function that is shared with `org-insert-company`.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>