fix(image): allow image pasting in Linux
This is really a massive hack, but it's something that I needed for
working inside of WSL. Most of the time when I'm in Emacs I'm on a Mac,
which means that `pngpaste` works great for inserting images. That's not
the case on Linux, which doesn't have an equivalent. This solution,
which is a stupid hack that I did just to get things out of the way,
makes use of https://github.com/cpbotha/save_image_from_clipboard to
save the image.

There's lots of optimizations that should be done here still. In
particular, it should do a better job of checking the binary to use,
only use on binary, remove hard coded paths, etc. But it gets me over
the hill for this point and allows me to save images that are grabbed
with Microsoft's Snipping Tool right into Emacs on WSL.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>